climatiqq-backend/requirements-docker.txt
Django==4.2.24
djangorestframework==3.15.2
django-cors-headers==4.7.0
djangorestframework-simplejwt==5.5.1
dj-database-url==3.0.1
psycopg2-binary==2.9.9
python-decouple==3.8
gunicorn==23.0.0
